+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlaygroundSettings.swift
@@ -307,12 +307,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
         case on
         case off
     }
-    enum InstantDebitsInDeferredIntents: String, PickerEnum {
-        static var enumName: String { "Instant Debits with Deferred Intents" }
-
-        case on
-        case off
-    }
     enum ExternalPaymentMethods: String, PickerEnum {
         static let enumName: String = "External PMs"
         // Based on https://git.corp.stripe.com/stripe-internal/stripe-js-v3/blob/55d7fd10/src/externalPaymentMethods/constants.ts#L13
@@ -475,7 +469,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
     var checkoutEndpoint: String
     var autoreload: Autoreload
     var shakeAmbiguousViews: ShakeAmbiguousViews
-    var instantDebitsInDeferredIntents: InstantDebitsInDeferredIntents
     var externalPaymentMethods: ExternalPaymentMethods
     var preferredNetworksEnabled: PreferredNetworksEnabled
     var requireCVCRecollection: RequireCVCRecollectionEnabled
@@ -522,7 +515,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
             checkoutEndpoint: Self.defaultCheckoutEndpoint,
             autoreload: .on,
             shakeAmbiguousViews: .off,
-            instantDebitsInDeferredIntents: .off,
             externalPaymentMethods: .off,
             preferredNetworksEnabled: .off,
             requireCVCRecollection: .off,